feat: restructure code
