+++ src/main/kotlin/com/likec4/jetbrainsplugin/lang/LikeC4File.kt
@@ -1,4 +1,6 @@
-package com.likec4.jetbrainsplugin
+package com.likec4.jetbrainsplugin.lang
+
+const val likeC4Name = "LikeC4"
 
 val supportedExtensions = setOf(
     "c4",
