+++ src/main/kotlin/com/likec4/jetbrainsplugin/lang/LikeC4FileType.kt
@@ -1,10 +1,11 @@
-package com.likec4.jetbrainsplugin
+package com.likec4.jetbrainsplugin.lang
 
 import com.intellij.openapi.fileTypes.LanguageFileType
+import com.likec4.jetbrainsplugin.icon.LikeC4Icons
 import javax.swing.Icon
 
 class LikeC4FileType : LanguageFileType(LikeC4Language) {
-    override fun getName() = "LikeC4"
+    override fun getName() = likeC4Name
     override fun getDefaultExtension() = "c4"
     override fun getIcon(): Icon = LikeC4Icons.LikeC4
     override fun getDescription() =
