+++ src/main/kotlin/com/likec4/jetbrainsplugin/lang/LikeC4Language.kt
@@ -1,0 +1,7 @@
+package com.likec4.jetbrainsplugin.lang
+
+import com.intellij.lang.Language
+
+object LikeC4Language : Language(likeC4Name) {
+    private fun readResolve(): Any = LikeC4Language
+}
