+++ src/main/kotlin/com/likec4/jetbrainsplugin/lsp/LikeC4LspServerSupportProvider.kt
@@ -1,4 +1,4 @@
-package com.likec4.jetbrainsplugin
+package com.likec4.jetbrainsplugin.lsp
 
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.vfs.VirtualFile
@@ -6,6 +6,8 @@ import com.intellij.platform.lsp.api.LspServer
 import com.intellij.platform.lsp.api.LspServerSupportProvider
 import com.intellij.platform.lsp.api.LspServerSupportProvider.LspServerStarter
 import com.intellij.platform.lsp.api.lsWidget.LspServerWidgetItem
+import com.likec4.jetbrainsplugin.icon.LikeC4Icons
+import com.likec4.jetbrainsplugin.lang.supportedExtensions
 
 class LikeC4LspServerSupportProvider : LspServerSupportProvider {
   override fun fileOpened(project: Project, file: VirtualFile, serverStarter: LspServerStarter) {
